@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.5+6
+version: 2.0.5+5
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -47,42 +49,41 @@ dependencies:
   retrofit: ^2.0.0
   logger: ^1.0.0
   fluttertoast: ^8.0.7
-#  country_code_picker: ^1.6.2
+  #  country_code_picker: ^1.6.2
   country_code_picker: ^2.0.2
- # progress_dialog: ^1.2.4
-#  sn_progress_dialog: ^1.0.2
+  # progress_dialog: ^1.2.4
+  #  sn_progress_dialog: ^1.0.2
   simple_fontellico_progress_dialog: ^0.2.0
   flutter_spinkit: ^5.0.0
   #without null safety
   cached_network_image: ^3.0.0
-#  cached_network_image: ^3.0.0
+  #  cached_network_image: ^3.0.0
   #old
-  #geocoder: ^0.2.1
+  # geocoder: ^0.2.1
   geocoding: ^2.0.0
   geolocator: ^7.2.0+1
   connectivity: ^3.0.6
   #without null safety
-#  flutter_widget_from_html: ^0.5.1+5
-#  flutter_widget_from_html: ^0.6.0
+  #  flutter_widget_from_html: ^0.5.1+5
+  #  flutter_widget_from_html: ^0.6.0
   flutter_html: ^2.1.0
   scoped_model: ^2.0.0-nullsafety.0
   razorpay_flutter: ^1.2.7
   stripe_payment: ^1.1.4
   sqflite: ^2.0.0+3
   path: ^1.8.0
-  onesignal_flutter: ^3.1.0
+  onesignal_flutter: ^3.2.3
   flutter_credit_card: ^2.0.0
   #modal_progress_hud: ^0.1.3
   modal_progress_hud_nsn: ^0.1.0-nullsafety-1
   flutter_multi_formatter: ^2.3.1
- # credit_card_number_validator: ^1.0.4
+  # credit_card_number_validator: ^1.0.4
   flutter_localizations:
     sdk: flutter
   http_auth: ^1.0.0
   #without null safety
   geocoder: ^0.2.1
-  #without null safety
-#  pull_to_refresh: ^1.6.3
+  #  pull_to_refresh: ^1.6.3
   pull_to_refresh: ^2.0.0
   webview_flutter: ^2.0.10
   url_launcher: ^6.0.9
