@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:mealup/utils/constants.dart';
+import 'package:homchf/utils/constants.dart';
 
 class HorizontalListViewWithImgText extends StatelessWidget {
 
