@@ -1,5 +1,5 @@
 
-import 'package:mealup/retrofit/server_error.dart';
+import 'package:homchf/retrofit/server_error.dart';
 
 class BaseModel<T> {
   late ServerError error;
