@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart' hide Headers;
-import 'package:mealup/utils/constants.dart';
+import 'package:homchf/utils/constants.dart';
 
 class ServerError implements Exception {
   int? _errorCode;
@@ -33,27 +33,11 @@ class ServerError implements Exception {
         break;
       case DioErrorType.response:
         _errorMessage = "Received invalid status code: ${error.response!.data}";
-        try {
-          if (error.response!.data['errors']['name'] != null) {
-            Constants.toastMessage(error.response!.data['errors']['name'][0]);
-            return;
-          } else if (error.response!.data['errors']['phone'] != null) {
-            Constants.toastMessage(error.response!.data['errors']['phone'][0]);
-            return;
-          } else if (error.response!.data['errors']['email_id'] != null) {
-            Constants.toastMessage(error.response!.data['errors']['email_id'][0]);
-            return;
-          } else if (error.response!.data['errors']['comment'] != null) {
-            Constants.toastMessage(error.response!.data['errors']['comment'][0]);
-            return;
-          } else {
-            Constants.toastMessage(error.response!.data['message'].toString());
-            return;
-          }
-        } catch (error1, stacktrace) {
+        try{
+          Constants.toastMessage(error.response!.data['message'].toString());
+        }catch (error1, stacktrace) {
           Constants.toastMessage(error.response!.data.toString());
-          print(
-              "Exception occurred: $error stackTrace: $stacktrace apiError: ${error.response!.data}");
+          print("Exception occurred: $error stackTrace: $stacktrace apiError: ${error.response!.data}"  );
         }
         break;
       case DioErrorType.cancel:
@@ -62,7 +46,7 @@ class ServerError implements Exception {
         break;
       case DioErrorType.other:
         _errorMessage =
-        "Connection failed. Please check your internet connection";
+        "Connection failed due to internet connection";
         Constants.toastMessage('Connection failed due to internet connection');
         break;
     }
