@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:mealup/screen_animation_utils/transitions.dart';
-import 'package:mealup/screens/bottom_navigation/dashboard_screen.dart';
-import 'package:mealup/utils/constants.dart';
-import 'package:mealup/utils/localization/language/languages.dart';
-import 'package:mealup/utils/preference_utils.dart';
+import 'package:homchf/screen_animation_utils/transitions.dart';
+import 'package:homchf/screens/bottom_navigation/dashboard_screen.dart';
+import 'package:homchf/utils/constants.dart';
+import 'package:homchf/utils/localization/language/languages.dart';
+import 'package:homchf/utils/preference_utils.dart';
 
 import 'intro_screen3.dart';
 
@@ -37,7 +37,7 @@ class IntroScreen2 extends StatelessWidget {
                       height: 40,),
                   ),
                 ),
-                Image.asset('images/ic_intro2.png'),
+                Image.asset('images/ic_intro2_1.png'),
                 Padding(
                   padding: const EdgeInsets.all(10.0),
                   child: Container(
@@ -49,7 +49,7 @@ class IntroScreen2 extends StatelessWidget {
                           textAlign: TextAlign.center,
                           style: TextStyle(
                             fontFamily: Constants.appFont,
-                            color: Constants.colorBlack,
+                            color: Color(Constants.colorBlack),
                             fontSize: 25.0,
                           ),
                         ),
@@ -59,7 +59,7 @@ class IntroScreen2 extends StatelessWidget {
                             Languages.of(context)!.labelScreenIntro2Line2,
                             textAlign: TextAlign.center,
                             style: TextStyle(
-                                color: Constants.colorGray,
+                                color: Color(Constants.colorGray),
                                 fontFamily: Constants.appFont,
                                 fontSize: 16.0),
                           ),
@@ -74,7 +74,7 @@ class IntroScreen2 extends StatelessWidget {
                     IconButton(
                       icon: Icon(
                         Icons.arrow_upward,
-                        color: Constants.colorTheme,
+                        color: Color(Constants.colorTheme),
                       ),
                       onPressed: () {
                         Navigator.pop(context);
@@ -98,7 +98,7 @@ class IntroScreen2 extends StatelessWidget {
                           Languages.of(context)!.labelSkip,
                           style: TextStyle(
                               letterSpacing: 3.0,
-                              color: Constants.colorTheme,
+                              color: Color(Constants.colorTheme),
                               fontFamily: Constants.appFont,
                               fontWeight: FontWeight.bold,
                               fontSize: 13.0),
@@ -108,7 +108,7 @@ class IntroScreen2 extends StatelessWidget {
                     IconButton(
                       icon: Icon(
                         Icons.arrow_downward,
-                        color: Constants.colorTheme,
+                        color: Color(Constants.colorTheme),
                       ),
                       onPressed: () {
                         Navigator.of(context).push(
