@@ -1,18 +1,28 @@
+import 'dart:typed_data';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:mealup/model/UserAddressListModel.dart';
-import 'package:mealup/retrofit/api_header.dart';
-import 'package:mealup/retrofit/api_client.dart';
-import 'package:mealup/retrofit/base_model.dart';
-import 'package:mealup/retrofit/server_error.dart';
-import 'package:mealup/screen_animation_utils/transitions.dart';
-import 'package:mealup/screens/bottom_navigation/dashboard_screen.dart';
-import 'package:mealup/utils/SharedPreferenceUtil.dart';
-import 'package:mealup/utils/app_toolbar.dart';
-import 'package:mealup/utils/constants.dart';
-import 'package:mealup/utils/localization/language/languages.dart';
+import 'package:geolocator/geolocator.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:homchf/model/UserAddressListModel.dart';
+import 'package:homchf/retrofit/api_header.dart';
+import 'package:homchf/retrofit/api_client.dart';
+import 'package:homchf/retrofit/base_model.dart';
+import 'package:homchf/retrofit/server_error.dart';
+import 'package:homchf/screen_animation_utils/transitions.dart';
+import 'package:homchf/screens/bottom_navigation/dashboard_screen.dart';
+import 'package:homchf/utils/SharedPreferenceUtil.dart';
+import 'package:homchf/utils/app_toolbar.dart';
+import 'package:homchf/utils/constants.dart';
+import 'package:homchf/utils/localization/language/languages.dart';
+import 'dart:ui' as ui;
 
+import 'package:homchf/screens/address/add_address_screen.dart';
+import 'package:homchf/screens/address/edit_address_screen.dart';
+import 'package:homchf/utils/app_toolbar_with_btn_clr.dart';
+import 'package:pull_to_refresh/pull_to_refresh.dart';
+import 'package:location/location.dart';
 
 class SetLocationScreen extends StatefulWidget {
   @override
@@ -27,14 +37,25 @@ class _SetLocationScreenState extends State<SetLocationScreen> {
   // String _city = '';
   // String _zipCode = '';
 
-
   List<UserAddressListData> _userAddressList = [];
+  RefreshController _refreshController = RefreshController(initialRefresh: false);
+  bool _isSyncing = false;
+  // Position currentLocation;
+  LocationData? _locationData;
+  Location _location = Location();
+  late Position currentLocation;
+  double? _currentLatitude = 0.0;
 
-  @override
-  void initState() {
-    super.initState();
-    Constants.checkNetwork().whenComplete(() => callGetUserAddresses());
-  }
+  double? _currentLongitude = 0.0;
+  BitmapDescriptor? _markerIcon;
+  
+  
+  
+  // @override
+  // void initState() {
+  //   super.initState();
+  //   Constants.checkNetwork().whenComplete(() => callGetUserAddresses());
+  // }
 
   Future<BaseModel<UserAddressListModel>> callGetUserAddresses() async {
     UserAddressListModel response;
@@ -60,6 +81,74 @@ class _SetLocationScreenState extends State<SetLocationScreen> {
     return BaseModel()..data = response;
   }
 
+  // getUserLocation() async {
+  //   currentLocation = await Geolocator.getCurrentPosition(desiredAccuracy: LocationAccuracy.high);
+  //   _currentLatitude = currentLocation.latitude;
+  //   _currentLongitude = currentLocation.longitude;
+  //   print('selectedLat $_currentLatitude');
+  //   print('selectedLng $_currentLongitude');
+  // }
+  Future<void> getUserLocation() async {
+    _locationData = await _location.getLocation();
+    if (_locationData != null) {
+      _currentLatitude = _locationData!.latitude;
+      _currentLongitude = _locationData!.longitude;
+      print('selectedLat $_currentLatitude');
+      print('selectedLng $_currentLongitude');
+    }
+  }
+void _onRefresh() async {
+    // monitor network fetch
+    await Future.delayed(Duration(milliseconds: 1000));
+    // if failed,use refreshFailed()
+    Constants.checkNetwork().whenComplete(() => callGetUserAddresses());
+    if (mounted) setState(() {});
+    _refreshController.refreshCompleted();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _createMarkerImageFromAsset(context);
+    Constants.checkNetwork().whenComplete(() => callGetUserAddresses());
+    getUserLocation();
+  }
+
+  Future<void> _createMarkerImageFromAsset(BuildContext context) async {
+    if (_markerIcon == null) {
+      BitmapDescriptor bitmapDescriptor =
+          await _bitmapDescriptorFromSvgAsset(context, 'images/ic_marker.svg');
+      //  _updateBitmap(bitmapDescriptor);
+      setState(() {
+        _markerIcon = bitmapDescriptor;
+      });
+    }
+  }
+
+  Future<BitmapDescriptor> _bitmapDescriptorFromSvgAsset(
+      BuildContext context, String assetName) async {
+    // Read SVG file as String
+    String svgString = await DefaultAssetBundle.of(context).loadString(assetName);
+    // Create DrawableRoot from SVG String
+    DrawableRoot svgDrawableRoot = await svg.fromSvgString(svgString, '');
+
+    // toPicture() and toImage() don't seem to be pixel ratio aware, so we calculate the actual sizes here
+    MediaQueryData queryData = MediaQuery.of(context);
+    double devicePixelRatio = queryData.devicePixelRatio;
+    double width = 32 * devicePixelRatio; // where 32 is your SVG's original width
+    double height = 32 * devicePixelRatio; // same thing
+
+    // Convert to ui.Picture
+    ui.Picture picture = svgDrawableRoot.toPicture(size: Size(width, height));
+
+    // Convert to ui.Image. toImage() takes width and height as parameters
+    // you need to find the best size to suit your needs and take into account the
+    // screen DPI
+    ui.Image image = await picture.toImage(width.toInt(), height.toInt());
+    ByteData? bytes = await (image.toByteData(format: ui.ImageByteFormat.png));
+    return BitmapDescriptor.fromBytes(bytes!.buffer.asUint8List());
+  }
+
   @override
   Widget build(BuildContext context) {
     dynamic screenWidth = MediaQuery.of(context).size.width;
@@ -74,8 +163,45 @@ class _SetLocationScreenState extends State<SetLocationScreen> {
 
     return SafeArea(
       child: Scaffold(
-        appBar: ApplicationToolbar(
+        appBar: ApplicationToolbarWithClrBtn(
           appbarTitle: Languages.of(context)!.labelSetLocation,
+          strButtonTitle: '+ ${Languages.of(context)!.labelAddAddress}',
+          btnColor: Color(Constants.colorThemeW),
+          onBtnPress: () {
+            if (_currentLongitude != 0.0) {
+              Navigator.pop(context);
+              Navigator.of(context).push(Transitions(
+                  transitionType: TransitionType.fade,
+                  curve: Curves.bounceInOut,
+                  reverseCurve: Curves.fastLinearToSlowEaseIn,
+                  // widget: HereMapDemo())
+                  widget: AddAddressScreen(
+                    isFromAddAddress: true,
+                    currentLat: _currentLatitude,
+                    currentLong: _currentLongitude,
+                    marker: _markerIcon,
+                  )));
+            }
+            else 
+            {
+              getUserLocation();
+              _currentLatitude =  currentLocation.latitude;
+              _currentLongitude = currentLocation.longitude;
+    
+              Navigator.pop(context);
+              Navigator.of(context).push(Transitions(
+                  transitionType: TransitionType.fade,
+                  curve: Curves.bounceInOut,
+                  reverseCurve: Curves.fastLinearToSlowEaseIn,
+                  // widget: HereMapDemo())
+                  widget: AddAddressScreen(
+                    isFromAddAddress: true,
+                    currentLat: _currentLatitude,
+                    currentLong: _currentLongitude,
+                    marker: _markerIcon,
+                  )));
+            }
+          },
         ),
         body: Container(
           margin: EdgeInsets.only(left: 20),
@@ -127,7 +253,7 @@ class _SetLocationScreenState extends State<SetLocationScreen> {
                                       style: TextStyle(
                                         fontSize: ScreenUtil().setSp(18),
                                         fontFamily: Constants.appFontBold,
-                                        color: Constants.colorTheme,
+                                        color: Color(Constants.colorTheme),
                                       ),
                                     ),
                                   )
@@ -186,7 +312,7 @@ class _SetLocationScreenState extends State<SetLocationScreen> {
                                           'images/ic_map.svg',
                                           width: 18,
                                           height: 18,
-                                          color: Constants.colorTheme,
+                                          color: Color(Constants.colorTheme),
                                         ),
                                         Expanded(
                                           child: Padding(
@@ -199,7 +325,8 @@ class _SetLocationScreenState extends State<SetLocationScreen> {
                                                   fontSize: 12,
                                                   fontFamily:
                                                       Constants.appFont,
-                                                  color: Constants.colorBlack),
+                                                  color: Color(Constants.colorBlack),
+                                              )
                                             ),
                                           ),
                                         )
