@@ -2,7 +2,6 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:mealup/utils/constants.dart';
 
 class MyWidget extends StatelessWidget {
   @override
@@ -56,7 +55,7 @@ class MyDynamicHeader extends SliverPersistentHeaderDelegate {
                 child: Center(
                   child: CircularProgressIndicator(
                     value: percentage,
-                    valueColor: AlwaysStoppedAnimation<Color>(Constants.colorWhite),
+                    valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
                   ),
                 )
             ),
