@@ -1,18 +1,21 @@
 import 'package:flutter/material.dart';
-import 'package:mealup/utils_google_map/place_service.dart';
+import 'package:homchf/utils_google_map/place_service.dart';
 // import 'package:google_api_headers/google_api_headers.dart';
 // import 'package:flutter_google_places/flutter_google_places.dart';
 // import 'package:google_maps_webservice/places.dart';
 // import 'package:geocoder/geocoder.dart';
 // import 'package:geocoder/services/base.dart';
 
+// const kGoogleApiKey = "AIzaSyCWtCJS-S1S8WLrke2jgEBXBl7qS3wL4rY";
 
+// GoogleMapsPlaces _places = GoogleMapsPlaces(apiKey: kGoogleApiKey);
 
 class AddressSearch extends SearchDelegate<Suggestion1> {
-  AddressSearch() {
-    apiClient = PlaceApiProvider();
+  AddressSearch(this.sessionToken) {
+    apiClient = PlaceApiProvider(sessionToken);
   }
 
+  final sessionToken;
   late PlaceApiProvider apiClient;
 
   @override
@@ -23,7 +26,7 @@ class AddressSearch extends SearchDelegate<Suggestion1> {
         icon: Icon(Icons.clear),
         onPressed: () {
           query = '';
-        },        
+        },
       )
     ];
   }
@@ -36,7 +39,7 @@ class AddressSearch extends SearchDelegate<Suggestion1> {
       tooltip: 'Back',
       icon: Icon(Icons.arrow_back),
       onPressed: () {
-        Suggestion1 suggestion1 = new Suggestion1('', '',0.0 ,0.0 );
+        Suggestion1 suggestion1 = new Suggestion1('', '', 0.0, 0.0);
         close(context, suggestion1);
       },
     );
@@ -45,13 +48,17 @@ class AddressSearch extends SearchDelegate<Suggestion1> {
   @override
   Widget buildResults(BuildContext context) {
     return FutureBuilder<List<Suggestion1>>(
-      future: query == "" ? null : apiClient.fetchLatlong(query, Localizations.localeOf(context).languageCode),
+      future: query == ""
+          ? null
+          : apiClient.fetchLatlong(
+              query, Localizations.localeOf(context).languageCode),
       builder: (context, snapshot) => query == ''
           ? Container(
               padding: EdgeInsets.all(16.0),
               child: Text('Enter your address'),
             )
-          : snapshot.hasData ? ListView.builder(
+          : snapshot.hasData
+              ? ListView.builder(
                   itemBuilder: (context, index) => ListTile(
                     title: Text((snapshot.data![index]).description),
                     onTap: () {
@@ -69,7 +76,8 @@ class AddressSearch extends SearchDelegate<Suggestion1> {
     return FutureBuilder<List<Suggestion1>>(
       future: query == ""
           ? null
-          : apiClient.fetchLatlong(query, Localizations.localeOf(context).languageCode),
+          : apiClient.fetchLatlong(
+              query, Localizations.localeOf(context).languageCode),
       builder: (context, snapshot) => query == ''
           ? Container(
               padding: EdgeInsets.all(16.0),
@@ -78,8 +86,7 @@ class AddressSearch extends SearchDelegate<Suggestion1> {
           : snapshot.hasData
               ? ListView.builder(
                   itemBuilder: (context, index) => ListTile(
-                    title:
-                        Text((snapshot.data![index]).description),
+                    title: Text((snapshot.data![index]).description),
                     onTap: () {
                       close(context, snapshot.data![index]);
                     },
